FIX: Pull external event info for past events in chapter page (#758)
previously, we werent loading this information when displaying past
events on the chapter timeline. in contrast, for the full events timeline,
we load all of the event information. because of this, the external
events associated with a chapter that are now displayed in the timeline
don't work as expected i.e. they dont redirect to the external url when
clicked.

this commit fixes this problem by simply loading the necessary info
from the database when passing the event info to the UI